requirements.txt
asgiref==3.4.1
astroid==2.11.5
autopep8==1.6.0
dill==0.3.4
Django==3.2
djangorestframework==3.13.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
Pillow==9.1.0
platformdirs==2.5.2
pycodestyle==2.8.0
pylint==2.13.8
pylint-django==2.5.3
pylint-plugin-utils==0.7
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
wrapt==1.14.1
